chore: add support for input-variables
Update tree is not working properly using just event-based... :-|

Too cumbersome to update the code. Consider using an observable
approach where the source of the data contains points (weakref) to
nodes iterested in the update (keep the from_slot, to_slot) around
to help differentiate the datasource. Similar to how JavaFX does
it.

Self-reference is not supported, circular dependencies may/may not
be supported but I'm not going to worry about that right now. This
is not even alpha anyway (aka somethinge for future me to worry about)